In this version, the maximum probability map is comprised of 142 brain regions in each hemisphere and available for the following coordinate spaces: + MNI ICBM 152 (2009c Nonlinear Asymmetric) + Colin 27 (1998, cropped)
In addition, this version complements the brain regions in the maximum probability map with 6 gap maps in each hemisphere to achieve full cortical coverage. Furthermore, this version includes surface mesh data for all cortical brain regions and gap maps for the Colin 27 (1998, cropped) coordinate space.

The maximum probability map of the brain regions and gap maps are stored in separate data files for each hemisphere and coordinate space. Surface mesh data are stored in separate data files for each hemisphere and brain region and gap map.